+++ spring-addons-starter-oidc/src/main/java/com/c4_soft/springaddons/security/oidc/starter/properties/condition/HasTokenEdpointParametersPropertiesCondition.java
@@ -2,7 +2,7 @@ package com.c4_soft.springaddons.security.oidc.starter.properties.condition;
 
 public class HasTokenEdpointParametersPropertiesCondition extends HasPropertyPrefixCondition {
 
-	public HasTokenEdpointParametersPropertiesCondition() {
-		super("com.c4-soft.springaddons.oidc.client.token-request-params");
-	}
+  public HasTokenEdpointParametersPropertiesCondition() {
+    super("com.c4-soft.springaddons.oidc.client.token");
+  }
 }