+++ spring-addons-starter-rest/src/main/java/com/c4_soft/springaddons/rest/SpringAddonsRestProperties.java
@@ -233,11 +233,6 @@ public class SpringAddonsRestProperties {
        */
       private Optional<Integer> readTimeoutMillis = Optional.empty();
 
-      /**
-       * Supported only with {@link RestClient}. Set the number of bytes to write in each chunk.
-       */
-      private Optional<Integer> chunkSize = Optional.empty();
-
       @Data
       public static class ProxyProperties {
         private boolean enabled = true;